chore: moved test files to tests folder
