+++ src/tests/components/PodcastCard.spec.tsx
@@ -1,33 +1,40 @@
 import React from "react";
 import { render, screen } from "@testing-library/react";
-import { PodcastCard, PodcastCardProps } from "./PodcastCard";
+import { PodcastCard } from "../../components/PodcastCard";
+import { BrowserRouter } from "react-router-dom";
 
 describe("test component rendering", () => {
-  let defaultProps: PodcastCardProps;
+  let renderComponent: JSX.Element;
   beforeAll(() => {
-    defaultProps = {
+    const defaultProps = {
       thumbnail: "",
       title: "My Title",
       author: "Author",
       id: "32552335",
     };
+
+    renderComponent = (
+      <BrowserRouter>
+        <PodcastCard {...defaultProps} />
+      </BrowserRouter>
+    );
   });
   it("renders title properly", () => {
-    render(<PodcastCard {...defaultProps} />);
+    render(renderComponent);
     const card = screen.getByRole("caption");
 
     expect(card).toHaveTextContent("My Title");
   });
 
   it("renders author properly", () => {
-    render(<PodcastCard {...defaultProps} />);
+    render(renderComponent);
     const authorDescription = screen.getByText(/Author/i);
 
     expect(authorDescription).toBeInTheDocument();
   });
 
   it("renders image with alt attribute", () => {
-    render(<PodcastCard {...defaultProps} />);
+    render(renderComponent);
     const image: HTMLImageElement = screen.getByAltText("My Title");
 
     expect(image.alt).toContain("My Title");
